refactoring in controller admin n album
