+++ app/helpers/index.js
@@ -1,11 +1,9 @@
-const constants = require('./constants');
 const utils = require('./utils');
 const regex = require('./regex');
 const statusCodes = require('./statusCodes');
 const errors = require('./errors');
 
 module.exports = {
-  ...constants,
   statusCodes,
   utils,
   regex,
