+++ test/controllers/signUpAdmin.test.js
@@ -8,36 +8,41 @@ describe('Register a administrator user', () => {
   beforeEach(async () => {
     const { UserService } = require('../../app/services');
     dataUser.signUpAdmin.password = '12345678';
+    dataUser.signUp.password = '12345678';
     await UserService.signUp(dataUser.signUpAdmin);
   });
 
   it('should sign up a administrator user', async done => {
-    const logInAdmin = await request(app)
+    const { token } = await request(app)
       .post('/users/sessions')
-      .send(dataUser.signInAdmin);
-    const { token } = JSON.parse(logInAdmin.text);
-    const res = await request(app)
+      .send(dataUser.signInAdmin)
+      .then(res => res.body);
+    await request(app)
       .post('/admin/users')
       .set('token', token)
-      .send(dataUser.signUp);
-    expect(res.statusCode).toBe(statusCodes.created);
-    done();
+      .send(dataUser.signUp)
+      .then(res => {
+        expect(res.statusCode).toBe(statusCodes.created);
+        done();
+      });
   });
 
   it('should update an existing user and grant the admin role', async done => {
     await request(app)
       .post('/users')
       .send(dataUser.signUp);
-    const logInAdmin = await request(app)
+    const { token } = await request(app)
       .post('/users/sessions')
-      .send(dataUser.signInAdmin);
-    const { token } = JSON.parse(logInAdmin.text);
-    const res = await request(app)
+      .send(dataUser.signInAdmin)
+      .then(res => res.body);
+    await request(app)
       .post('/admin/users')
       .set('token', token)
-      .send(dataUser.signUp);
-    expect(res.statusCode).toBe(statusCodes.created);
-    done();
+      .send(dataUser.signUp)
+      .then(res => {
+        expect(res.statusCode).toBe(statusCodes.created);
+        done();
+      });
   });
 
   it('should reject to created for permissions', async done => {
@@ -66,10 +71,10 @@ describe('Register a administrator user', () => {
   });
 
   it('should reject for data user empty', async done => {
-    const logInAdmin = await request(app)
+    const { token } = await request(app)
       .post('/users/sessions')
-      .send(dataUser.signInAdmin);
-    const { token } = JSON.parse(logInAdmin.text);
+      .send(dataUser.signInAdmin)
+      .then(res => res.body);
     const res = await request(app)
       .post('/admin/users')
       .set('token', token);
